+++ lib/services/decup_service.dart
@@ -20,6 +20,7 @@ abstract interface class DecupService {
   static const int ack = 0x1C;
   static const int nak = 0xFC;
   int? get closeCode;
+  String? get closeReason;
   Future<void> get ready;
   Stream<Uint8List> get stream;
   Future close([int? closeCode, String? closeReason]);