+++ lib/services/fake_zusi_service.dart
@@ -27,6 +27,9 @@ class FakeZusiService implements ZusiService {
   int? get closeCode => _controller.isClosed ? 1005 : null;
 
   @override
+  String? get closeReason => closeCode != null ? 'Timeout' : null;
+
+  @override
   Future<void> get ready => Future.delayed(const Duration(seconds: 1));
 
   @override