+++ lib/services/ws_decup_service.dart
@@ -35,6 +35,9 @@ class WsDecupService implements DecupService {
   int? get closeCode => _channel.closeCode;
 
   @override
+  String? get closeReason => closeCode != null ? 'Timeout' : null;
+
+  @override
   Future<void> get ready => _channel.ready;
 
   @override