+++ lib/services/ws_ota_service.dart
@@ -30,6 +30,9 @@ class WsOtaService implements OtaService {
   int? get closeCode => _channel.closeCode;
 
   @override
+  String? get closeReason => closeCode != null ? 'Timeout' : null;
+
+  @override
   Future<void> get ready => _channel.ready;
 
   @override