+++ lib/widgets/mdu_dialog.dart
@@ -162,9 +162,9 @@ class _MduDialogState extends ConsumerState<MduDialog> {
     }
 
     final msg = await _retryOnFailure(() => _mdu.configTransferRate(0));
-    if (msg.contains(MduService.ack)) {
+    if (msg.contains(MduService.ack) || _mdu.closeReason != null) {
       _updateEphemeralState(
-        status: 'No common transfer rate found',
+        status: _mdu.closeReason ?? 'No common transfer rate found',
         progress: 0,
       );
       return Uint8List.fromList([MduService.ack, MduService.ack]);
@@ -179,8 +179,11 @@ class _MduDialogState extends ConsumerState<MduDialog> {
     final msg = await _retryOnFailure(
       () => _mdu.zppValidQuery(widget._zpp!.id, widget._zpp!.flash.length),
     );
-    if (msg.contains(MduService.ack)) {
-      _updateEphemeralState(status: 'ZPP not valid', progress: 0);
+    if (msg.contains(MduService.ack) || _mdu.closeReason != null) {
+      _updateEphemeralState(
+        status: _mdu.closeReason ?? 'ZPP not valid',
+        progress: 0,
+      );
       return Uint8List.fromList([MduService.ack, MduService.ack]);
     }
     return msg;
@@ -195,8 +198,11 @@ class _MduDialogState extends ConsumerState<MduDialog> {
     final msg = await _retryOnFailure(
       () => _mdu.zppLcDcQuery(widget._zpp!.developerCode),
     );
-    if (msg.contains(MduService.ack)) {
-      _updateEphemeralState(status: 'Load code not valid', progress: 0);
+    if (msg.contains(MduService.ack) || _mdu.closeReason != null) {
+      _updateEphemeralState(
+        status: _mdu.closeReason ?? 'Load code not valid',
+        progress: 0,
+      );
       return Uint8List.fromList([MduService.ack, MduService.ack]);
     }
     return msg;
@@ -208,8 +214,11 @@ class _MduDialogState extends ConsumerState<MduDialog> {
     var msg = await _retryOnFailure(
       () => _mdu.zppErase(0, widget._zpp!.flash.length),
     );
-    if (msg.contains(MduService.ack)) {
-      _updateEphemeralState(status: 'Erasing failed', progress: 0);
+    if (msg.contains(MduService.ack) || _mdu.closeReason != null) {
+      _updateEphemeralState(
+        status: _mdu.closeReason ?? 'Erasing failed',
+        progress: 0,
+      );
       return Uint8List.fromList([MduService.ack, MduService.ack]);
     }
 
@@ -220,16 +229,22 @@ class _MduDialogState extends ConsumerState<MduDialog> {
       msg = await _retryOnFailure(() => _mdu.busy());
       if (!msg.contains(MduService.ack)) break;
     }
-    if (msg.contains(MduService.ack)) {
-      _updateEphemeralState(status: 'Erasing failed', progress: 0);
+    if (msg.contains(MduService.ack) || _mdu.closeReason != null) {
+      _updateEphemeralState(
+        status: _mdu.closeReason ?? 'Erasing failed',
+        progress: 0,
+      );
       return Uint8List.fromList([MduService.ack, MduService.ack]);
     }
 
     //
     _updateEphemeralState(status: 'Writing');
     msg = await _zppWrite(widget._zpp!.flash);
-    if (msg.contains(MduService.ack)) {
-      _updateEphemeralState(status: 'Writing failed', progress: 0);
+    if (msg.contains(MduService.ack) || _mdu.closeReason != null) {
+      _updateEphemeralState(
+        status: _mdu.closeReason ?? 'Writing failed',
+        progress: 0,
+      );
       return Uint8List.fromList([MduService.ack, MduService.ack]);
     }
 
@@ -237,8 +252,11 @@ class _MduDialogState extends ConsumerState<MduDialog> {
     msg = await _retryOnFailure(
       () => _mdu.zppUpdateEnd(0, widget._zpp!.flash.length),
     );
-    if (msg.contains(MduService.ack)) {
-      _updateEphemeralState(status: 'ZPP update end check failed', progress: 0);
+    if (msg.contains(MduService.ack) || _mdu.closeReason != null) {
+      _updateEphemeralState(
+        status: _mdu.closeReason ?? 'ZPP update end check failed',
+        progress: 0,
+      );
       return Uint8List.fromList([MduService.ack, MduService.ack]);
     }
 
@@ -278,6 +296,11 @@ class _MduDialogState extends ConsumerState<MduDialog> {
         }
       }
 
+      // WebSocket closed on the server side
+      if (_mdu.closeReason != null) {
+        return Uint8List.fromList([MduService.ack, MduService.ack]);
+      }
+
       // Update progress
       _updateEphemeralState(
         status:
@@ -310,12 +333,15 @@ class _MduDialogState extends ConsumerState<MduDialog> {
         final match = exp.firstMatch(entry.value.name);
         if (match != null && match[0] != null) {
           final String name = match[0]!;
-          setState(() {
-            _decoders[decoderId] = ListTile(
-              leading: const Icon(Icons.circle),
-              title: Text(name),
-            );
-          });
+          _updateEphemeralState(
+            decoder: MapEntry(
+              decoderId,
+              ListTile(
+                leading: const Icon(Icons.circle),
+                title: Text(name),
+              ),
+            ),
+          );
         }
       }
     }
@@ -331,20 +357,25 @@ class _MduDialogState extends ConsumerState<MduDialog> {
   /// \todo document
   Future<Uint8List> _zsuUpdate(int decoderId) async {
     // Ping all decoders
-    _updateEphemeralState(status: 'Ping');
-    setState(() {
-      _decoders[decoderId] = ListTile(
-        leading: const Icon(Icons.pending),
-        title: _decoders[decoderId]!.title,
-      );
-    });
+    _updateEphemeralState(
+      status: 'Ping',
+      decoder: MapEntry(
+        decoderId,
+        ListTile(
+          leading: const Icon(Icons.pending),
+          title: _decoders[decoderId]!.title,
+        ),
+      ),
+    );
     await _retryOnFailure(() => _mdu.ping(0, 0));
 
     // Ping decoder to update
     var msg = await _retryOnFailure(() => _mdu.ping(0, decoderId));
-    if (msg[0] == MduService.ack || msg[1] == MduService.nak) {
+    if (msg[0] == MduService.ack ||
+        msg[1] == MduService.nak ||
+        _mdu.closeReason != null) {
       _updateEphemeralState(
-        status: 'Decoder does not respond',
+        status: _mdu.closeReason ?? 'Decoder does not respond',
         progress: 0,
         decoder: MapEntry(
           decoderId,
@@ -367,9 +398,9 @@ class _MduDialogState extends ConsumerState<MduDialog> {
     msg = await _retryOnFailure(
       () => _mdu.zsuErase(0, zsuFirmware.bin.length - 1),
     );
-    if (msg.contains(MduService.ack)) {
+    if (msg.contains(MduService.ack) || _mdu.closeReason != null) {
       _updateEphemeralState(
-        status: 'Erasing failed',
+        status: _mdu.closeReason ?? 'Erasing failed',
         progress: 0,
         decoder: MapEntry(
           decoderId,
@@ -389,17 +420,20 @@ class _MduDialogState extends ConsumerState<MduDialog> {
     }
 
     // Write flash
-    _updateEphemeralState(status: 'Writing');
-    setState(() {
-      _decoders[decoderId] = ListTile(
-        leading: const Icon(Icons.download_for_offline),
-        title: _decoders[decoderId]!.title,
-      );
-    });
+    _updateEphemeralState(
+      status: 'Writing',
+      decoder: MapEntry(
+        decoderId,
+        ListTile(
+          leading: const Icon(Icons.download_for_offline),
+          title: _decoders[decoderId]!.title,
+        ),
+      ),
+    );
     msg = await _zsuWrite(zsuFirmware.bin);
-    if (msg.contains(MduService.ack)) {
+    if (msg.contains(MduService.ack) || _mdu.closeReason != null) {
       _updateEphemeralState(
-        status: 'Writing failed',
+        status: _mdu.closeReason ?? 'Writing failed',
         progress: 0,
         decoder: MapEntry(
           decoderId,
@@ -420,9 +454,9 @@ class _MduDialogState extends ConsumerState<MduDialog> {
         crc32(zsuFirmware.bin),
       ),
     );
-    if (msg.contains(MduService.ack)) {
+    if (msg.contains(MduService.ack) || _mdu.closeReason != null) {
       _updateEphemeralState(
-        status: 'ZSU update CRC32 check failed',
+        status: _mdu.closeReason ?? 'ZSU update CRC32 check failed',
         progress: 0,
         decoder: MapEntry(
           decoderId,
@@ -437,9 +471,9 @@ class _MduDialogState extends ConsumerState<MduDialog> {
 
     // CRC32 result
     msg = await _retryOnFailure(() => _mdu.zsuCrc32Result());
-    if (msg.contains(MduService.ack)) {
+    if (msg.contains(MduService.ack) || _mdu.closeReason != null) {
       _updateEphemeralState(
-        status: 'ZSU update CRC32 check failed',
+        status: _mdu.closeReason ?? 'ZSU update CRC32 check failed',
         progress: 0,
         decoder: MapEntry(
           decoderId,
@@ -497,6 +531,11 @@ class _MduDialogState extends ConsumerState<MduDialog> {
         else {
           return msg;
         }
+      }
+
+      // WebSocket closed on the server side
+      if (_mdu.closeReason != null) {
+        return Uint8List.fromList([MduService.ack, MduService.ack]);
       }
 
       // Update progress