+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - App close confirmation ([#58](https://github.com/OpenRemise/Frontend/issues/58))
 - ZIMO update URLs are hardcoded ([#66](https://github.com/OpenRemise/Frontend/issues/66))
 - Animate between `NavigationRail` and `NavigationBar` ([#52](https://github.com/OpenRemise/Frontend/issues/52))
+- Bugfix detect WebSocket closed on the server side ([#64](https://github.com/OpenRemise/Frontend/issues/64))
 
 ## 0.0.8
 - Add favicon ([#56](https://github.com/OpenRemise/Frontend/issues/56))